backend/requirements.txt
# Core
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
gunicorn>=21.2.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-multipart>=0.0.9

# Database
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
psycopg[binary]>=3.2.0
alembic>=1.13.0
pgvector>=0.2.5

# Caching & Queue
redis>=5.0.0
celery>=5.3.0

# Auth
passlib[bcrypt]>=1.7.4
pyjwt>=2.8.0

# AI
google-genai>=1.0.0
sentence-transformers>=3.0.0

# Knowledge Ingestion (RAG)
pypdf>=4.0.0
pdfplumber>=0.11.0
pytesseract>=0.3.10
pdf2image>=1.17.0

# External APIs
httpx>=0.27.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
lxml>=5.1.0

# Storage
boto3>=1.34.0
python-magic>=0.4.27

# Image Processing
Pillow>=10.2.0

# Observability
sentry-sdk[fastapi]>=2.0.0

# Utilities
python-dotenv>=1.0.0
structlog>=24.1.0
orjson>=3.9.0

# SSE
sse-starlette>=1.6.0
